ref: rename and split plugins.base tests so they actually run
