+++ tests/sentry/plugins/base/response_test.py
@@ -1,0 +1,11 @@
+from sentry.plugins.base.response import JSONResponse
+
+
+def test_json_response():
+    resp = JSONResponse({}).respond(None)
+    assert resp.status_code == 200
+
+
+def test_json_response_with_status_kwarg():
+    resp = JSONResponse({}, status=400).respond(None)
+    assert resp.status_code == 400
